% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLogbookVarsLikeEffort2PopAnnuallndic.R
\name{addLogbookVarsLikeEffort2PopAnnualIndic}
\alias{addLogbookVarsLikeEffort2PopAnnualIndic}
\title{Link annual indic to effort}
\usage{
addLogbookVarsLikeEffort2PopAnnualIndic(general, a_variable = "effort",
  nby = 6)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
Link annual indic to effort
}
\examples{
\dontrun{
general <- setGeneralOverallVariable(main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )

 
 
loadLoglikeFiles (general, use_port_info=FALSE)


 

displace_annual_indic_per_pop_year <- addLogbookVarsLikeEffort2PopAnnualIndic(general)


 

 }
}
