% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotTotLandingsPerSce.R
\name{barplotTotLandingsPerSce}
\alias{barplotTotLandingsPerSce}
\title{Making an overview barplot for total landings per scenario for a selected set of stocks}
\usage{
barplotTotLandingsPerSce(general = general, type_of_column = "pop",
  selected = "_selected_set1_", selected_pops = c(0, 1, 2, 3, 11, 23,
  24, 26, 30, 31, 32), group1 = c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31,
  32), selected_scenarios = general$namefolderoutput,
  scenarios_names = general$namefolderoutput, nby = 5,
  firsty = "2015", lasty = "2019", a_width = 3500, a_height = 2000,
  black_and_white = TRUE, ylims = c(0, 10))
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function processes further the lst_loglike objects
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )



  loadLoglikeFiles(general=general, use_port_info=FALSE)



  barplotTotLandingsPerSce (general=general,
                                   selected="_selected_set1_",
                                   type_of_column="pop",
                                   selected_pops=c(0:38),
                                   group1= c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31, 32),
                                   selected_scenarios=general$namefolderoutput,
                                   scenarios_names=general$namefolderoutput,
                                   nby=5, firsty="2015", lasty="2019",
                                   a_width=3500, a_height=2000, black_and_white = FALSE, ylims = c(0,305))

 Note that if black_and_white at true then the barplot is split into 2 groups
(first group is group1 stocks, second is all other ones)


barplotTotLandingsPerSce (general=general,
                                   type_of_column="disc",
                                   selected="_selected_set1_",
                                   selected_pops=c(0:38),
                                   group1= c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31, 32),
                                   selected_scenarios=general$namefolderoutput,
                                   scenarios_names=general$namefolderoutput,
                                   nby=5, firsty="2015", lasty="2019",
                                   a_width=3500, a_height=2000, black_and_white = FALSE, ylims = c(0,305))


  }
}
