% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCompareSimSimPlots.R
\name{callCompareSimSimPlots}
\alias{callCompareSimSimPlots}
\title{Call compareSimSimPlot to generate a wealth of graphics (better to call callCompareSimSimPlot())}
\usage{
callCompareSimSimPlots(general, the_baseline = "svana_baseline",
  explicit_pops = explicit_pops, selected = "_selected_set1_",
  selected_vessels = selected_set1, years_span = 2015:2019,
  per_pop = FALSE, per_vessel = TRUE)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function generates numerous graphics in a hierarchy of folders
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (pathToRawInputs =file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_gis_", 
                                                 general$case_study)),
                                      pathToDisplaceInputs = file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_", general$case_study)),
                                      pathToOutputs =file.path("C:","DISPLACE_outputs"),
                                      caseStudy="DanishFleet",
                                      iGraph=41,
                                      iYear="2015",
                                      iCountry="DEN",
                                      nbPops=39,
                                      nbSzgroup=14,
                                      nameFolderInput="DanishFleet",
                                      theScenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbSimus=20,
                                      useSQLite=FALSE    
                                      )




callCompareSimSimPlots (general,
                          the_baseline="svana_baseline",
                          explicit_pops=explicit_pops,
                          selected = "_selected_set1_",
                          selected_vessels=selected_set1,
                          years_span=2015:2019,
                            per_pop=FALSE,
                          per_vessel=TRUE)
 #=> produce files in output folders....
 }
}
