% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePieplotStressPerPort.R
\name{comparePieplotStressPerPort}
\alias{comparePieplotStressPerPort}
\title{Produce pie plots centered on ports with proportion affected vessels in classes of stress}
\usage{
comparePieplotStressPerPort(general, a_variable = "rev_from_av_prices",
  nbsim = 20, the_baseline = "svana_baseline",
  a_polygon_including_interesting_ports = list(x = c(-2.850909,
  16.955191, 18.233003, -12.434505), y = c(46.47831, 52.22793, 63.08831,
  62.44946)), selected_scenarios = c("svana_sub1mx20", "svana_sub4mx20"),
  name_scenarios = c("svana_sub1mx20", "svana_sub4mx20"),
  gis_shape = list(svana_sub1mx20 = list(NSsub1mx20, BSsub1mx20),
  svana_sub4mx20 = list(NSsub4mx20, BSsub4mx20)), a_width = 3400,
  a_height = 3500, xlims = c(-1, 17), ylims = c(53, 60))
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function produces pieplots centered on ports of %  of vessel affected per bin for various indicators facing baseline sce
}
\examples{
\dontrun{
general <- setGeneralOverallVariable (pathToRawInputs =file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_gis_", 
                                                 general$case_study)),
                                      pathToDisplaceInputs = file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_", general$case_study)),
                                      pathToOutputs =file.path("C:","DISPLACE_outputs"),
                                      caseStudy="DanishFleet",
                                      iGraph=41,
                                      iYear="2015",
                                      iCountry="DEN",
                                      nbPops=39,
                                      nbSzgroup=14,
                                      nameFolderInput="DanishFleet",
                                      theScenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbSimus=20,
                                      useSQLite=FALSE    
                                      )




  library(maptools)
  NSsub1mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                    'DISPLACE_SVANAProject', 'Input for DISPLACE', 'NST2_sub1_mx20_wgs84'), proj4string=CRS("+proj=longlat +ellps=WGS84"))
  BSsub1mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                    'DISPLACE_SVANAProject', 'Input for DISPLACE', 'BHT2_Sub1_Mx_20_wgs84'), proj4string=CRS("+proj=longlat +ellps=WGS84"))
  NSsub4mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                      'DISPLACE_SVANAProject', 'Input for DISPLACE', 'NST2_sub4_mx_20_wgs84'), proj4string=CRS("+proj=longlat +ellps=WGS84"))
  BSsub4mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                        'DISPLACE_SVANAProject', 'Input for DISPLACE', 'BHT2_Sub4_Mx_20LongTailedD_wgs84'), proj4string=CRS("+proj=longlat #'+ellps=WGS84"))


 
 
loadLoglikeFiles (general, use_port_info=TRUE)


comparePieplotStressPerPort (general,
                                      nbsim=20,
                                      the_baseline="svana_baseline",
                                      a_polygon_including_interesting_ports= list(x=c(-2.850909,  16.955191,  18.233003, -12.434505),
                                                                                  y=c(46.47831, 52.22793, 63.08831, 62.44946)),
                                      selected_scenarios=  c("svana_sub1mx20","svana_sub4mx20"), 
                                      gis_shape=list(
                                          svana_sub1mx20=   list(NSsub1mx20, BSsub1mx20),
                                          svana_sub4mx20=   list(NSsub4mx20, BSsub4mx20)),
                                          a_width= 3200, a_height =2100, xlims =  c(7, 16), ylims = c(53.5,59)
                                          ) 
 }
}
