% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSimSimIndividualCumTraj.R
\name{compareSimSimIndividualVidCumTraj}
\alias{compareSimSimIndividualVidCumTraj}
\title{Generate individual vessels time series on output variables from vessel aggregations}
\usage{
compareSimSimIndividualVidCumTraj(general = general,
  vessel_indexes = 200:250, a_var = "gradva",
  sce_baseline = "svana_baseline", sce_tested = "svana_sub4mx20",
  nbiters = 1, nby = 5)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function is a quick way to help interpretation of results by a quick and dirty check on individual trajectories
}
\examples{
                                                         

\dontrun{
general <- setGeneralOverallVariable (main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )


 loadLoglikeFiles (general, use_port_info=FALSE)



 compareSimSimIndividualVidCumTraj(general=general,
                                    vessel_indexes=200:210,
                                    a_var="gradva", 
                                    sce_baseline="svana_baseline", 
                                    sce_tested="svana_sub4mx20",
                                    nbiters=1,
                                    nby=5)

}
}
