% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/expressAggLoglikeFilesIndicatorsRelativeToBaselineSce.R
\name{expressAggLoglikeFilesIndicatorsRelativeToBaselineSce}
\alias{expressAggLoglikeFilesIndicatorsRelativeToBaselineSce}
\title{Produce a table relative to the baseline scenario}
\usage{
expressAggLoglikeFilesIndicatorsRelativeToBaselineSce(general,
  the_baseline = "svana_baseline", sets = c("_selected_set1_",
  "_selected_set2_", "_selected_set3_"))
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function produce a table of indicator from processing further the aggregate loglike data
}
\examples{
\dontrun{
general <- setGeneralOverallVariable (pathToRawInputs =file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_gis_", 
                                                 "DanishFleet")),
                                      pathToDisplaceInputs = file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_", "DanishFleet")),
                                      pathToOutputs =file.path("C:","DISPLACE_outputs"),
                                      caseStudy="DanishFleet",
                                      iGraph=41,
                                      iYear="2015",
                                      iCountry="DEN",
                                      nbPops=39,
                                      nbSzgroup=14,
                                      theScenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbSimus=20,
                                      useSQLite=FALSE    
                                      )



if(FALSE){
 getAggLoglikeFiles(general=general, what="weight",
            explicit_pops=c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31, 32),
            implicit_pops=c (4, 5, 6, 7, 8, 9, 10, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 25, 27, 28, 29, 33, 34, 35, 36, 37, 38),
            selected_vessels_set1=selected_vessels_set_1,
            selected_vessels_set2=selected_vessels_set_2,
            selected_vessels_set3=selected_vessels_set_3)
} else{
  loadLoglikeFiles (general, use_port_info=FALSE)
}


expressAggLoglikeFilesIndicatorsRelativeToBaselineSce(general,
                             the_baseline = "svana_baseline",
                             sets=c("_selected_set1_", "_selected_set2_", "_selected_set3_"))


boxplotAggLoglikeFilesIndicators (general= general,
                                            the_baseline="svana_baseline",
                                            sets=c("_selected_set1_", "_selected_set2_", "_selected_set3_"))

  }
}
