% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAggNodesLayerFiles.R
\name{getAggNodeLayerFiles}
\alias{getAggNodeLayerFiles}
\title{Produce an average spatial layer as first step to do a map}
\usage{
getAggNodeLayerFiles(general, a_type = "cumcatches", a_tstep = "34321")
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function generates an average layer as a first step to generate maps from popnodes files
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (pathToRawInputs =file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_gis_", 
                                                 general$case_study)),
                                      pathToDisplaceInputs = file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_", general$case_study)),
                                      pathToOutputs =file.path("C:","DISPLACE_outputs"),
                                      caseStudy="DanishFleet",
                                      iGraph=41,
                                      iYear="2015",
                                      iCountry="DEN",
                                      nbPops=39,
                                      nbSzgroup=14,
                                      nameFolderInput="DanishFleet",
                                      theScenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbSimus=20,
                                      useSQLite=FALSE    
                                      )




getAggNodeLayerFiles (general, a_type="cumcatches", a_tstep="34321")
 #=> produce files in output folders....
 }
}
