% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAggPoplikeFiles.R
\name{getAggPoplikeFiles}
\alias{getAggPoplikeFiles}
\title{Processing the popdyn.dat files}
\usage{
getAggPoplikeFiles(general = general, explicit_pops = c(0, 1, 2, 3, 11,
  23, 24, 26, 30, 31, 32), the_baseline = "svana_baseline")
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function process the popdyn_simu_xxx.dat files storing Ns at pop
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline", 
                                                      "svana_sub1mx20", 
                                                      "svana_sub4mx20", 
                                                      "svana_sub4mx5ns20bt", 
                                                      "svana_sub4mx20ns5bt", 
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )





  getAggPoplikeFiles(general=general,
                     explicit_pops=c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31, 32),
                     the_baseline ="svana_baseline"
                     )  

  }
}
