% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodeAverageLayerFiles.R
\name{mapNodeAverageLayerFiles}
\alias{mapNodeAverageLayerFiles}
\title{Generate maps from averaging stochastic DISPLACE spatial layera}
\usage{
mapNodeAverageLayerFiles(general, a_type = "cumcatches",
  the_baseline = "svana_baseline",
  selected_scenarios_for_plot = general$namefolderoutput,
  selected_scenarios_for_table = general$namefolderoutput,
  selected_areas_for_table = c("22", "23", "24", "25", "IIIa", "IVa", "IVb",
  "IVc"), the_breaks_baseline = c(0.5, 1, round(exp(seq(0.5, 14, by = 1.2))),
  1e+06), gis_shape = list(), a_width = 3400, a_height = 3500,
  xlims = c(-1, 17), ylims = c(53, 60), xcell = 12, ycell = 17)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function generates maps from an average layer as a second step after call to getAggNodesLayerFiles()
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )


  # caution: could take a  while...
  getAggNodeLayerFiles (general, a_type="cumcatches", a_tstep="34321")


  library(maptools)
  sh_coastlines               <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub','DISPLACE_input_myfish','graphsspe', 'shp', 'francois_EU'), 
                                                      proj4string=CRS("+proj=longlat +ellps=WGS84"))
  ices_areas                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub','DISPLACE_input_raw','ices_areas','ices_areas'), 
                                                     proj4string=CRS("+proj=longlat +ellps=WGS84"))

  NSsub1mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                    'DISPLACE_SVANAProject', 'Input for DISPLACE', 'NST2_sub1_mx20_wgs84'),
                                                     proj4string=CRS("+proj=longlat +ellps=WGS84"))
  BSsub1mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                    'DISPLACE_SVANAProject', 'Input for DISPLACE', 'BHT2_Sub1_Mx_20_wgs84'), 
                                                       proj4string=CRS("+proj=longlat +ellps=WGS84"))
  NSsub4mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                      'DISPLACE_SVANAProject', 'Input for DISPLACE', 'NST2_sub4_mx_20_wgs84'), 
                                                       proj4string=CRS("+proj=longlat +ellps=WGS84"))
  BSsub4mx20                  <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                        'DISPLACE_SVANAProject', 'Input for DISPLACE', 'BHT2_Sub4_Mx_20LongTailedD_wgs84'), 
                                                          proj4string=CRS("+proj=longlat #'+ellps=WGS84"))
  NSsub4mx5                   <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                         'DISPLACE_SVANAProject', 'Input for DISPLACE', 'NST2_sub4_mx05_wgs84'), 
                                                           proj4string=CRS("+proj=longlat +ellps=WGS84"))
  BSsub4mx5                   <- readShapePoly(file.path('C:','Users','fbas','Documents','GitHub',
                                                          'DISPLACE_SVANAProject', 'Input for DISPLACE', 'BHT2_Sub4_Mx_5LgtailedD_wgs84'), 
                                                             proj4string=CRS("+proj=longlat +ellps=WGS84"))



  mapNodeAverageLayerFiles (general, a_type="cumcatches",   the_baseline= "svana_baseline",
                           selected_scenarios_for_plot=general$namefolderoutput,
                           selected_scenarios_for_table=general$namefolderoutput,
                           selected_areas_for_table=c("22",    "23",    "24",    "25",    "IIIa",  "IVa",   "IVb",   "IVc"),
                           the_breaks_baseline= c(0.5, 1, round(exp(seq(0.5, 14, by=1.1))), 1000000),
                           gis_shape=list(svana_baseline=   list(sh_coastlines), # ices_areas),
                                          svana_sub1mx20=   list(NSsub1mx20, BSsub1mx20),
                                          svana_sub4mx20=   list(NSsub4mx20, BSsub4mx20),
                                          svana_sub4mx5ns20bt=   list(NSsub4mx5, BSsub4mx20),
                                          svana_sub4mx20ns5bt=   list(NSsub4mx20, BSsub4mx5),
                                          svana_sub4mx5ns5bt=    list(NSsub4mx5, BSsub4mx5),
                                          a_width= 3400, a_height =3500, xlims =  c(-1, 17), ylims = c(53,60), xcell=12, ycell=17
                                          ))
}
}
