% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonPlotsFromAnnualIndicFiles.R
\name{polygonPlotsFromAnnualIndicFiles}
\alias{polygonPlotsFromAnnualIndicFiles}
\title{Produce polygon plots of time series for annual indicators}
\usage{
polygonPlotsFromAnnualIndicFiles(general = general,
  a_variable = "Fbar", nby = 5, the_baseline = "svana_baseline",
  a_width = 3500, a_height = 1000,
  selected_scenarios = general$namefolderoutput[1:3],
  the_scenario_names = general$namefolderoutput[1:3],
  explicit_pops = c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31, 32),
  is_individual_export = TRUE, add_legend = FALSE)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function produces all accumulated time series over month to compare scenario outcomes
All the plots are being stored in a polygons_plots folder that can further be found in the output folder.
(compare up to 5 scenarios simultaneously)
}
\examples{
\dontrun{
general <- setGeneralOverallVariable(main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )


loadLoglikeFiles(general=general, use_port_info=FALSE)



polygonPlotsFromAnnualIndicFiles (general=general,
                                           a_variable="Fbar",  
                                           nby=5,
                                           the_baseline="svana_baseline",
                                           a_width=3500,
                                           a_height=1000,
                                           selected_scenarios=general$namefolderoutput[1:3],
                                           the_scenario_names=general$namefolderoutput[1:3], 
                                           explicit_pops=c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31, 32),
                                           is_individual_export=TRUE,
                                           add_legend=FALSE
                                           )

  }
}
