% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneralVariable.R
\name{setGeneralOverallVariable}
\alias{setGeneralOverallVariable}
\title{Setting the general variable to use across other routines}
\usage{
setGeneralOverallVariable(main_path_outputs = file.path("C:",
  "DISPLACE_outputs"), case_study = "DanishFleet", igraph = 41,
  a.year = "2015", a.country = "DEN", nbpops = 39, nbszgroup = 14,
  namefolderinput = "DanishFleet", the_scenarios = c("svana_baseline",
  "svana_sub1mx20", "svana_sub4mx20", "svana_sub4mx5ns20bt",
  "svana_sub4mx20ns5bt", "svana_sub4mx5ns5bt"), nbsimus = 20)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function sets the general variable useful for relative file paths etc.
}
\examples{
general <- setGeneralOverallVariable()
}
