% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stressBarplot.R
\name{stressBarplot}
\alias{stressBarplot}
\title{stress barplots in percentage of affected vessels}
\usage{
stressBarplot(general = general, the_baseline = "svana_baseline",
  selected_vessels = selected_vessels_set_1, by_class = NULL,
  a_width = 3000, a_height = 2000)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function produces barplots of percentage of vessels affected per bin for various indicators facing baseline sce
}
\examples{
if(FALSE){

 


 general <- setGeneralOverallVariable(main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )

  selected_vessels_set_1 <- as.character(read.table(file.path(general$main.path, general$case_study,
          paste("selected_vessels_set_1.dat",sep='')), header=FALSE)[,1])


 # CAUTION: this following process can take a while!
 compareSimSimPlots(general=general, lst_loglike_agg_1=lst_loglike_agg_1, lst_loglike_agg_2=lst_loglike_agg_2, years_span=2015:2019, ...,
                                  explicit_pops=explicit_pops, plot_obs=TRUE,
                                      idx.sim=list(sce1=c(1), sce2=c(1)), combined_name=c("baseline_vs_implicit"),
                                        a.comment="", what="per_vessel", what2="weight", count=0,
                                         a.xlab="", a.ylab="", a.unit=1, do_mtext=FALSE)


  require(plotrix)
  stressBarplot (general=general,
                         the_baseline="svana_baseline",
                         selected_vessels=selected_vessels_set_1,
                         by_class=NULL,
                         a_width=1500, a_height=3500)


 }   # end FALSE
}
