% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLogbookVarsLikeEffort2PopAnnuallndic.R
\name{addLogbookVarsLikeEffort2PopAnnualIndic}
\alias{addLogbookVarsLikeEffort2PopAnnualIndic}
\title{Link annual indic to effort}
\usage{
addLogbookVarsLikeEffort2PopAnnualIndic(general, a_variable = "effort",
  nby = 6)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
Link annual indic to effort
}
\examples{
\dontrun{
general <- setGeneralOverallVariable (pathToRawInputs =file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_gis_", 
                                                 general$case_study)),
                                      pathToDisplaceInputs = file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_", general$case_study)),
                                      pathToOutputs =file.path("C:","DISPLACE_outputs"),
                                      caseStudy="DanishFleet",
                                      iGraph=41,
                                      iYear="2015",
                                      iCountry="DEN",
                                      nbPops=39,
                                      nbSzgroup=14,
                                      nameFolderInput="DanishFleet",
                                      theScenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbSimus=20,
                                      useSQLite=FALSE    
                                      )



 
 
loadLoglikeFiles (general, use_port_info=FALSE)


 

displace_annual_indic_per_pop_year <- addLogbookVarsLikeEffort2PopAnnualIndic(general)


 

 }
}
