% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCompareSimSimPlots.R
\name{callCompareSimSimPlots}
\alias{callCompareSimSimPlots}
\title{Call compareSimSimPlot to generate a wealth of graphics (better to call callCompareSimSimPlot())}
\usage{
callCompareSimSimPlots(general, the_baseline = "svana_baseline",
  explicit_pops = explicit_pops, selected = "_selected_set1_",
  selected_vessels = selected_set1, years_span = 2015:2019,
  per_pop = FALSE, per_vessel = TRUE)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function generates numerous graphics in a hierarchy of folders
}
\examples{
general <- setGeneralOverallVariable (main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )


callCompareSimSimPlots (general,
                          the_baseline="svana_baseline",
                          explicit_pops=explicit_pops,
                          selected = "_selected_set1_",
                          selected_vessels=selected_set1,
                          years_span=2015:2019,
                            per_pop=FALSE,
                          per_vessel=TRUE)
 #=> produce files in output folders....

}
