% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/expressAggLoglikeFilesIndicatorsRelativeToBaselineSce.R
\name{expressAggLoglikeFilesIndicatorsRelativeToBaselineSce}
\alias{expressAggLoglikeFilesIndicatorsRelativeToBaselineSce}
\title{Produce a table relative to the baseline scenario}
\usage{
expressAggLoglikeFilesIndicatorsRelativeToBaselineSce(general,
  the_baseline = "svana_baseline", sets = c("_selected_set1_",
  "_selected_set2_", "_selected_set3_"))
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function produce a table of indicator from processing further the aggregate loglike data
}
\examples{
general <- setGeneralOverallVariable (main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )


if(FALSE){
 aggregratLoglikeFiles(general=general, what="weight",
            explicit_pops=explicit_pops2,
            implicit_pops=c (4, 5, 6, 7, 8, 9, 10, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 25, 27, 28, 29, 33, 34, 35, 36, 37, 38),
            selected_vessels_set1=selected_vessels_set_1,
            selected_vessels_set2=selected_vessels_set_2,
            selected_vessels_set3=selected_vessels_set_3)
} else{
  loadLoglikeFiles (general, use_port_info=FALSE)
}


expressAggLoglikeFilesIndicatorsRelativeToBaselineSce(general,
                             the_baseline = "svana_baseline",
                             sets=c("_selected_set1_", "_selected_set2_", "_selected_set3_"))


boxplotAggLoglikeFilesIndicators (general= general,
                                            the_baseline="svana_baseline",
                                            sets=c("_selected_set1_", "_selected_set2_", "_selected_set3_"))


}
