% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAggNodesBenthosLayerFiles.R
\name{getAggNodeBenthosLayerFiles}
\alias{getAggNodeBenthosLayerFiles}
\title{Produce an average spatial layer as first step to do a map}
\usage{
getAggNodeBenthosLayerFiles(general, a_tstep = "34321")
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function generates an average layer as a first step to generate maps from popnodes files 
Two different types of files are going to be written behind the scene i.e. one file storing the stack of all input files and keeping the sce dimension, the other doing the averaging per sce.
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (main_path_outputs =file.path("C:","DISPLACE_outputs"),
                                      case_study="DanishFleet",
                                      igraph=41,
                                      a.year="2015",
                                      a.country="DEN",
                                      nbpops=39,
                                      nbszgroup=14,
                                      namefolderinput="DanishFleet",
                                      the_scenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbsimus=20
                                      )


getAggNodeBenthosLayerFiles (general,  a_tstep="34321")
 #=> produce files in output folders....
 }
}
