% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAggPoplikeFiles.R
\name{getAggPoplikeFiles}
\alias{getAggPoplikeFiles}
\title{Processing the popdyn.dat files}
\usage{
getAggPoplikeFiles(general = general, explicit_pops = c(0, 1, 2, 3, 11,
  23, 24, 26, 30, 31, 32), the_baseline = "svana_baseline")
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function process the popdyn_simu_xxx.dat files storing Ns at pop
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (pathToRawInputs =file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_gis_", 
                                                 "DanishFleet")),
                                      pathToDisplaceInputs = file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_", "DanishFleet")),
                                      pathToOutputs =file.path("C:","DISPLACE_outputs"),
                                      caseStudy="DanishFleet",
                                      iGraph=41,
                                      iYear="2015",
                                      iCountry="DEN",
                                      nbPops=39,
                                      nbSzgroup=14,
                                      theScenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbSimus=20,
                                      useSQLite=FALSE    
                                      )






  getAggPoplikeFiles(general=general,
                     explicit_pops=c(0, 1, 2, 3, 11, 23, 24, 26, 30, 31, 32),
                     the_baseline ="svana_baseline"
                     )  

  }
}
