% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAggLoglikeFiles.R
\name{loadLoglikeFiles}
\alias{loadLoglikeFiles}
\title{Re-load outcome files of getAggLoglikeFiles() in the global env.}
\usage{
loadLoglikeFiles(general = general, use_port_info = FALSE)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function re-load the files produced by getAggLoglikeFiles()
}
\examples{
 \dontrun{
loadLoglikeFiles(general=general, use_port_info=FALSE)
}
}
