% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonPlotsFromAggNodesBenthosFiles.R
\name{polygonPlotsFromAggNodeBenthosFiles}
\alias{polygonPlotsFromAggNodeBenthosFiles}
\title{Produce polygons time series plots for Benthos indicators and per functional groups}
\usage{
polygonPlotsFromAggNodeBenthosFiles(general, is_relative = TRUE,
  the_baseline = "svana_baseline")
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function generates polygon plots in absolute or relative to the baseline scenario.
Input data files are searched in the folder hierarchy.
Only the median is displayed if in relative terms because the uncertainty range cannot be handled in this case
}
\examples{

\dontrun{
general <- setGeneralOverallVariable (pathToRawInputs =file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_gis_", 
                                                 "DanishFleet")),
                                      pathToDisplaceInputs = file.path("C:", "Users", "fbas", 
                                                "Documents", "GitHub", paste0("DISPLACE_input_", "DanishFleet")),
                                      pathToOutputs =file.path("C:","DISPLACE_outputs"),
                                      caseStudy="DanishFleet",
                                      iGraph=41,
                                      iYear="2015",
                                      iCountry="DEN",
                                      nbPops=39,
                                      nbSzgroup=14,
                                      theScenarios= c("svana_baseline",
                                                      "svana_sub1mx20",
                                                      "svana_sub4mx20",
                                                      "svana_sub4mx5ns20bt",
                                                      "svana_sub4mx20ns5bt",
                                                      "svana_sub4mx5ns5bt" ),
                                      nbSimus=20,
                                      useSQLite=FALSE    
                                      )


getAggNodeBenthosLayerFiles (general,  a_tstep="34321")
 #=> produce files in output folders....


  polygonPlotsFromAggNodeBenthosFiles (general, 
                                               is_relative= TRUE,
                                               the_baseline= "svana_baseline")


 }
}
