% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneralVariable.R
\name{setGeneralOverallVariable}
\alias{setGeneralOverallVariable}
\title{Setting the general variable to use across other routines}
\usage{
setGeneralOverallVariable(pathToRawInputs = file.path("C:", "Users",
  "fbas", "Documents", "GitHub", paste0("DISPLACE_input_gis_",
  general$case_study)), pathToDisplaceInputs = file.path("C:", "Users",
  "fbas", "Documents", "GitHub", paste0("DISPLACE_input_",
  general$case_study)), pathToOutputs = file.path("C:",
  "DISPLACE_outputs"), caseStudy = "DanishFleet", iGraph = 41,
  iYear = "2015", iCountry = "DEN", nbPops = 39, nbSzgroup = 14,
  nameFolderInput = "DanishFleet", theScenarios = c("svana_baseline",
  "svana_sub1mx20", "svana_sub4mx20", "svana_sub4mx5ns20bt",
  "svana_sub4mx20ns5bt", "svana_sub4mx5ns5bt"), nbSimus = 20,
  useSQLite = FALSE)
}
\arguments{
\item{fname}{First name}

\item{lname}{Last name}
}
\description{
This function sets the general variable useful for relative file paths etc.
}
\examples{
general <- setGeneralOverallVariable()
}
